StreetCLIP/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
pandas>=1.2.0
numpy>=1.19.0
tqdm>=4.50.0
clip @ git+https://github.com/openai/CLIP.git
matplotlib>=3.4.0 